Add files via upload
Project 1: It is first project which is madlib game. It takes user's input for adjective, noun, name, and verb and later use on existing paragraph. Here we play with the strings.
Project 2: This project is about guessing the number randomly generated by computer. Here , library function of python random is imported.
Project 3: In this project we let computer guess our number. Random function is imported here too as in project 2.